react/pyproject.toml
[tool.poetry]
name = "iot-react-3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-anthropic = "^0.1.20"
langchain = "^0.2.7"
langgraph = "^0.1.8"
langchain-openai = "^0.1.16"
langchainhub = "^0.1.20"
black = "^24.4.2"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.7"
boto3 = "^1.34.144"
langchain-google-genai = "^1.0.7"
langchain-fireworks = "^0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
